Fixed the WiFi stack crashing after a few seconds
The base station broadcasts a precise byte in UDP to let the modules
know its IP address. However, the ESP was still listening to them, but
not handling them.
My theory is that the UDP buffer fills up, totally stucking the WiFi
stack (the ESP did not even respond to pings !)